Improve handling of missing/corrupted port file

+++ lib/cylc/scheduler.py
@@ -106,7 +106,6 @@ class Scheduler(object):
     EVENT_STALLED = 'stalled'
 
     # Intervals in seconds
-    INTERVAL_FS_CHECK = 600.0
     INTERVAL_MAIN_LOOP = 1.0
     INTERVAL_STOP_KILL = 10.0
     INTERVAL_STOP_PROCESS_POOL_EMPTY = 0.5
@@ -1573,10 +1572,22 @@ To see if %(suite)s is running on '%(host)s:%(port)s':
                 self.check_suite_stalled()
             now = time()
             if time_next_fs_check is None or now > time_next_fs_check:
-                if os.path.exists(suite_run_dir):
-                    time_next_fs_check = now + self.INTERVAL_FS_CHECK
-                else:
-                    raise SchedulerError("Suite run directory not found")
+                for message, item in [
+                        ("%s: suite run directory not found", suite_run_dir),
+                        ("%s: port file not found", self.port_file)]:
+                    if not os.path.exists(item):
+                        raise SchedulerError(message % (item))
+                try:
+                    port, host = open(self.port_file).read().splitlines()
+                    assert self.port == int(port) and self.host == host
+                except (AssertionError, IOError, IndexError, ValueError):
+                    exc = SchedulerError(
+                        ("%s: port file corrupted/modified and" +
+                         " will be left") % self.port_file)
+                    self.port_file = None
+                    raise exc
+                time_next_fs_check = (
+                    now + self._get_cylc_conf('health check interval'))
 
             if self.options.profile_mode:
                 now = time()
@@ -1967,7 +1978,7 @@ To see if %(suite)s is running on '%(host)s:%(port)s':
         sys.stdout.flush()
         sys.stderr.flush()
 
-        if self.comms_daemon:
+        if self.comms_daemon and self.port_file:
             try:
                 os.unlink(self.port_file)
             except OSError as exc:
@@ -2169,6 +2180,18 @@ To see if %(suite)s is running on '%(host)s:%(port)s':
             self.log.warning("Cannot get CPU % statistics: %s" % exc)
             return
         self._update_profile_info("CPU %", cpu_frac, amount_format="%.1f")
+
+    def _get_cylc_conf(self, key, default=None):
+        """Return a named setting under [cylc] from suite.rc or global.rc."""
+        for getter in [self.config.cfg['cylc'], GLOBAL_CFG.get(['cylc'])]:
+            try:
+                value = getter[key]
+            except KeyError:
+                pass
+            else:
+                if value is not None:
+                    return value
+        return default
 
     def _get_events_conf(self, key, default=None):
         """Return a named [cylc][[events]] configuration."""
